# typed storage