--- a/FreeAPS/Resources/javascript/prepare/middleware.js
+++ b/FreeAPS/Resources/javascript/prepare/middleware.js
@@ -3,7 +3,7 @@
 *     middleware_fn: String,
 *     glucose: [GlucoseEntry0],
 *     current_temp: TempBasal,
-*     iob: [IOBItem],
+*     iob: [IOBEntry],
 *     profile: Profile,
 *     autosens: Autosens?,
 *     meal: RecentCarbs,