--- a/FreeAPS/Sources/APS/OpenAPS/OpenAPS.swift
+++ b/FreeAPS/Sources/APS/OpenAPS/OpenAPS.swift
@@ -39,6 +39,7 @@ final class OpenAPS {
                         let tempBasal = currentTemp
                         self.storage.tempBasal.save(tempBasal)
 
+                        now = Date.now
                         let (
                             pumpHistory,
                             carbs,