--- a/FreeAPS/Sources/APS/Storage/AnnouncementsStorage.swift
+++ b/FreeAPS/Sources/APS/Storage/AnnouncementsStorage.swift
@@ -24,19 +24,18 @@ final class BaseAnnouncementsStorage: AnnouncementsStorage, Injectable {
 
     func storeAnnouncements(_ announcements: [Announcement], enacted: Bool) {
         processQueue.sync {
-            let file = enacted ? OpenAPS.FreeAPS.announcementsEnacted : OpenAPS.FreeAPS.announcements
-            self.storage.transaction { storage in
-                storage.append(announcements, to: file, uniqBy: \.createdAt)
-                let uniqEvents = storage.retrieve(file, as: [Announcement].self)?
+            let announcementStorage = enacted ? self.storage.announcementsEnacted : self.storage.announcements
+            self.storage.transaction { _ in
+                let uniqEvents = announcementStorage.append(announcements, uniqBy: \.createdAt)
                     .filter { $0.createdAt.addingTimeInterval(1.days.timeInterval) > Date() }
-                    .sorted { $0.createdAt > $1.createdAt } ?? []
-                storage.save(Array(uniqEvents), as: file)
+                    .sorted { $0.createdAt > $1.createdAt }
+                announcementStorage.save(Array(uniqEvents))
             }
         }
     }
 
     func syncDate() -> Date {
-        guard let events = storage.retrieve(OpenAPS.FreeAPS.announcementsEnacted, as: [Announcement].self),
+        guard let events = storage.announcementsEnacted.retrieveOpt(),
               let recentEnacted = events.filter({ $0.enteredBy == Announcement.remote }).first
         else {
             return Date().addingTimeInterval(-Config.recentInterval)
@@ -45,7 +44,7 @@ final class BaseAnnouncementsStorage: AnnouncementsStorage, Injectable {
     }
 
     func recent() -> Announcement? {
-        guard let events = storage.retrieve(OpenAPS.FreeAPS.announcements, as: [Announcement].self)
+        guard let events = storage.announcements.retrieveOpt()
         else {
             return nil
         }
@@ -57,7 +56,7 @@ final class BaseAnnouncementsStorage: AnnouncementsStorage, Injectable {
         else {
             return nil
         }
-        guard let enactedEvents = storage.retrieve(OpenAPS.FreeAPS.announcementsEnacted, as: [Announcement].self)
+        guard let enactedEvents = storage.announcementsEnacted.retrieveOpt()
         else {
             return recent
         }
@@ -70,7 +69,7 @@ final class BaseAnnouncementsStorage: AnnouncementsStorage, Injectable {
     }
 
     func recentEnacted() -> Announcement? {
-        guard let enactedEvents = storage.retrieve(OpenAPS.FreeAPS.announcementsEnacted, as: [Announcement].self)
+        guard let enactedEvents = storage.announcementsEnacted.retrieveOpt()
         else {
             return nil
         }
@@ -83,7 +82,7 @@ final class BaseAnnouncementsStorage: AnnouncementsStorage, Injectable {
     }
 
     func validate() -> [Announcement] {
-        guard let enactedEvents = storage.retrieve(OpenAPS.FreeAPS.announcementsEnacted, as: [Announcement].self)?.reversed()
+        guard let enactedEvents = storage.announcementsEnacted.retrieveOpt()?.reversed()
         else {
             return []
         }