… FreeAPS/Sources/APS/Storage/CoreDataStorage.swift
… FreeAPS/Sources/APS/Storage/GlucoseStorage.swift
… FreeAPS/Sources/APS/Storage/PumpHistoryStorage.swift
… FreeAPS/Sources/APS/Storage/TempTargetsStorage.swift
… FreeAPS/Sources/Helpers/Formatters.swift
… FreeAPS/Sources/Helpers/JSON.swift
… FreeAPS/Sources/Models/AutoIsfInput.swift
… FreeAPS/Sources/Models/BGTargets.swift
… FreeAPS/Sources/Models/CarbRatios.swift
… FreeAPS/Sources/Models/DetermineBasalInput.swift
… FreeAPS/Sources/Models/IOBItem.swift
… FreeAPS/Sources/Models/InsulinSensitivities.swift
… FreeAPS/Sources/Models/MiddlewareInput.swift
… FreeAPS/Sources/Models/PumpSettings.swift
… FreeAPS/Sources/Modules/AddCarbs/AddCarbsProvider.swift
… FreeAPS/Sources/Modules/AutotuneConfig/AutotuneConfigProvider.swift
… FreeAPS/Sources/Modules/AutotuneConfig/AutotuneConfigStateModel.swift
… FreeAPS/Sources/Modules/BasalProfileEditor/BasalProfileEditorProvider.swift
… FreeAPS/Sources/Modules/Bolus/BolusProvider.swift
… FreeAPS/Sources/Modules/CREditor/CREditorProvider.swift
… FreeAPS/Sources/Modules/ConfigEditor/ConfigEditorDataFlow.swift
… FreeAPS/Sources/Modules/ConfigEditor/ConfigEditorProvider.swift
… FreeAPS/Sources/Modules/ConfigEditor/ConfigEditorStateModel.swift
… FreeAPS/Sources/Modules/ConfigEditor/View/ConfigEditorRootView.swift
… FreeAPS/Sources/Modules/DataTable/DataTableProvider.swift
… FreeAPS/Sources/Modules/Dynamic/DynamicStateModel.swift
… FreeAPS/Sources/Modules/Home/HomeProvider.swift
… FreeAPS/Sources/Modules/Home/HomeStateModel.swift
… FreeAPS/Sources/Modules/ISFEditor/ISFEditorProvider.swift
… FreeAPS/Sources/Modules/NightscoutConfig/NightscoutConfigStateModel.swift