Introducing a typed `EntityStorage<T>` class (with a couple of extensions - for defaults and appendable lists).

In `FileStorage`, each file now has a corresponding entity storage, defined like the following:

```swift
    // the simplest - will read from disk, fall back to defaults json, can return nil
    let tempBasal = EntityStorage<TempBasal>(
        file: OpenAPS.Monitor.tempBasal,
        codec: EntityCodecs.json(),
    )

    // same as above + fall back value, has a method that never returns nil (no .append)
    let iob = EntityStorageWithFallback<[IOBEntry]>(
        file: OpenAPS.Monitor.iob,
        codec: EntityCodecs.json(),
        fallbackValue: []
    )

    // a variant for array types, with the .append methods (which behave the same as before)
    let carbsHistory = EntityStorageWithAppend<CarbsEntry>(
        file: OpenAPS.Monitor.carbHistory,
    )

    // decimals
    let reservoir = EntityStorage<Decimal>(
        file: OpenAPS.Monitor.reservoir,
        codec: EntityCodecs.decimal,
    )

    // raw string
    let middleware = EntityStorage<String>(
        file: OpenAPS.Middleware.determineBasal,
        codec: EntityCodecs.string,
        readDefaults: true,
    )
```

---

With these storages in place, the whole codebase has been updated to use them, instead of raw `storage.retrieve` / `storage.save`: no need to provide the filename every time, or the defaults, or the type.

Examples:

```swift
storage.pumpHistory.retrieveOrEmpty()
self.storage.iob.save(iob)

uniqEvents = storage.carbsHistory.append(futureCarbArray, uniqBy: \.id) // append now returns the updated array, no need to read again afterwards
   // ... filters
storage.carbsHistory.save(Array(uniqEvents))
```

Another bonus - each store ensures that the same type is read and written from the file.
* [ ] there is one TODO where the types of data written and read from one of the files are different

---

Another optimization - in-memory caching. Whenever an `EntityStorage` reads or writes, it memorizes the data in-memory and re-uses it later on.

Even better would be to memorize the decoded data, but since it's not all immutable - it is not safe, unless we refactor the models to be immutable.

---

While testing this, I noticed that "Loading files" for determine basal was still taking 50-60 milliseconds in the simulator, 30-40 milliseconds of which were taking to load the glucose.json (only 280 records). This didn't make any sense.

It turned out that the custom Date deserialization strategy was not great. I changed the way it's specified and now loading files for determine basal takes only 5 milliseconds.

```
before:
time for retrieve monitor/glucose.json: 30.161023139953613 ms
...
Time for Loading files 32.865047454833984 ms

after:
Time for Loading files 0.004907011985778809 seconds
```

---

Updated the code editor screens to work with this, since we're no longer working with raw files and filenames.

One of the runs in the simulator:
```
MakeProfiles: Time for Loading files 0.0010880231857299805 seconds
Time for TDD: 0.007158041000366211 seconds
dynamicVariables: Time for fetchTempTargetsSlider 0.00022089481353759766 seconds, total: 0.0011720657348632812
dynamicVariables: Time for fetchNumberOfOverrides 0.00032007694244384766 seconds, total: 0.0014960765838623047
dynamicVariables: Time for fetchTempTargets 9.500980377197266e-05 seconds, total: 0.0015950202941894531
Time for tdd and DynamicVariables 0.0857919454574585 seconds, total: 0.08689892292022705
MakeProfiles: Time for profile and pumpProfile 0.2605910301208496 seconds, total: 0.347540020942688
MakeProfiles: Time for save files 0.004487037658691406 seconds, total: 0.35318493843078613
Time for Loading files 0.004907011985778809 seconds
Time for tdd 0.001109004020690918 seconds
Time for Meal and IOB module 0.029888033866882324 seconds, total: 0.037515997886657715
Time for Determine Basal module 0.02254796028137207 seconds, total: 0.07565999031066895
```